data_server/kafka_crypto_producer/requirements.txt
binance==0.3
binance_connector==1.18.0
kafka-python==2.0.2